refactor(ios): fold DeviceCtlClient open-URL primitive into DeviceAppInspector (#3099)
Retire the parallel devicectl wrapper DeviceCtlClient and re-home its
open-URL primitive onto DeviceAppInspector, which is already the devicectl
wrapper (install/uninstall/copy/launch/terminate) and carries the darwin +
host-control gates.

- Move the DeviceUrlLauncher seam + isAvailable/launchWithPayloadUrl onto
  DeviceAppInspector; behavior byte-identical (same command string, same
  error text, same host-control posture).
- Extract a single private isHostControlMode() helper and collapse the six
  duplicated shouldUseHostControl() && isRunningInDocker() expressions.
- Drop the duplicated quoteShell / defaultDependencies (one copy remains).
- Rewire OpenURL injection + FakeDeviceCtlClient import; migrate the
  DeviceCtlClient unit tests onto the DeviceAppInspector method.

Closes #3099.
